+++ sympy/simplify/__init__.py
@@ -4,10 +4,9 @@ for example:
 the expression E**(pi*I) will be converted into -1
 the expression (x+x)**2 will be converted into 4*x**2
 """
-from .simplify import (ratsimp,
-    simplify, hypersimp, hypersimilar,
-    logcombine, separatevars, posify, polarify,
-    unpolarify, signsimp, ratsimpmodprime, bottom_up, nsimplify)
+from .simplify import (simplify, hypersimp, hypersimilar,
+    logcombine, separatevars, posify,
+    signsimp, bottom_up, nsimplify)
 
 from .fu import FU, fu
 
@@ -30,3 +29,5 @@ from .besselsimp import besselsimp
 from .powsimp import powsimp, powdenest
 
 from .combsimp import combsimp
+
+from .ratsimp import ratsimp, ratsimpmodprime
