+++ sympy/simplify/simplify.py
@@ -8,7 +8,7 @@ from sympy.core import (Basic, S, Add, Mul, Pow,
     FunctionClass, symbols,
     expand_power_exp)
 from sympy.core.add import _unevaluated_Add
-from sympy.core.compatibility import (combinations_with_replacement, iterable,
+from sympy.core.compatibility import (iterable,
     reduce, default_sort_key,
     ordered, range, as_int)
 from sympy.core.numbers import Float, I, pi, Rational, Integer
@@ -18,7 +18,7 @@ from sympy.core.rules import Transform
 from sympy.core.evaluate import global_evaluate
 from sympy.functions import (
     gamma, exp, sqrt, log, root, exp_polar,
-    sin, cos, tan, cosh, piecewise_fold)
+    sin, piecewise_fold)
 from sympy.functions.elementary.complexes import polar_lift, principal_branch
 from sympy.functions.elementary.exponential import ExpBase
 from sympy.functions.elementary.hyperbolic import HyperbolicFunction
@@ -28,9 +28,6 @@ from sympy.functions.combinatorial.factorials import CombinatorialFunction
 
 from sympy.utilities.iterables import has_variety
 
-from sympy.utilities.misc import debug
-
-
 from sympy.simplify.radsimp import radsimp, collect_sqrt, fraction
 from sympy.simplify.trigsimp import trigsimp, exptrigsimp
 from sympy.simplify.besselsimp import besselsimp
@@ -39,10 +36,9 @@ from sympy.simplify.cse_opts import sub_pre, sub_post
 from sympy.simplify.sqrtdenest import sqrtdenest
 from sympy.simplify.combsimp import combsimp
 
-from sympy.polys import (Poly, together, reduced, cancel, factor,
-    ComputationFailed, lcm, gcd, parallel_poly_from_expr)
-from sympy.polys.monomials import Monomial, monomial_div
-from sympy.polys.polyerrors import PolificationFailed, DomainError
+from sympy.polys import (together, cancel, factor,
+    lcm, gcd)
+from sympy.polys.polyerrors import DomainError
 
 
 import mpmath
@@ -203,363 +199,6 @@ def _separatevars_dict(expr, symbols):
         ret[k] = Mul(*v)
 
     return ret
-
-
-def ratsimp(expr):
-    """
-    Put an expression over a common denominator, cancel and reduce.
-
-    Examples
-    ========
-
-    >>> from sympy import ratsimp
-    >>> from sympy.abc import x, y
-    >>> ratsimp(1/x + 1/y)
-    (x + y)/(x*y)
-    """
-
-    f, g = cancel(expr).as_numer_denom()
-    try:
-        Q, r = reduced(f, [g], field=True, expand=False)
-    except ComputationFailed:
-        return f/g
-
-    return Add(*Q) + cancel(r/g)
-
-
-def ratsimpmodprime(expr, G, *gens, **args):
-    """
-    Simplifies a rational expression ``expr`` modulo the prime ideal
-    generated by ``G``.  ``G`` should be a Groebner basis of the
-    ideal.
-
-    >>> from sympy.simplify.simplify import ratsimpmodprime
-    >>> from sympy.abc import x, y
-    >>> eq = (x + y**5 + y)/(x - y)
-    >>> ratsimpmodprime(eq, [x*y**5 - x - y], x, y, order='lex')
-    (x**2 + x*y + x + y)/(x**2 - x*y)
-
-    If ``polynomial`` is False, the algorithm computes a rational
-    simplification which minimizes the sum of the total degrees of
-    the numerator and the denominator.
-
-    If ``polynomial`` is True, this function just brings numerator and
-    denominator into a canonical form. This is much faster, but has
-    potentially worse results.
-
-    References
-    ==========
-
-    M. Monagan, R. Pearce, Rational Simplification Modulo a Polynomial
-    Ideal,
-    http://citeseer.ist.psu.edu/viewdoc/summary?doi=10.1.1.163.6984
-    (specifically, the second algorithm)
-    """
-    from sympy import solve
-
-    quick = args.pop('quick', True)
-    polynomial = args.pop('polynomial', False)
-    debug('ratsimpmodprime', expr)
-
-    # usual preparation of polynomials:
-
-    num, denom = cancel(expr).as_numer_denom()
-
-    try:
-        polys, opt = parallel_poly_from_expr([num, denom] + G, *gens, **args)
-    except PolificationFailed:
-        return expr
-
-    domain = opt.domain
-
-    if domain.has_assoc_Field:
-        opt.domain = domain.get_field()
-    else:
-        raise DomainError(
-            "can't compute rational simplification over %s" % domain)
-
-    # compute only once
-    leading_monomials = [g.LM(opt.order) for g in polys[2:]]
-    tested = set()
-
-    def staircase(n):
-        """
-        Compute all monomials with degree less than ``n`` that are
-        not divisible by any element of ``leading_monomials``.
-        """
-        if n == 0:
-            return [1]
-        S = []
-        for mi in combinations_with_replacement(range(len(opt.gens)), n):
-            m = [0]*len(opt.gens)
-            for i in mi:
-                m[i] += 1
-            if all([monomial_div(m, lmg) is None for lmg in
-                    leading_monomials]):
-                S.append(m)
-
-        return [Monomial(s).as_expr(*opt.gens) for s in S] + staircase(n - 1)
-
-    def _ratsimpmodprime(a, b, allsol, N=0, D=0):
-        """
-        Computes a rational simplification of ``a/b`` which minimizes
-        the sum of the total degrees of the numerator and the denominator.
-
-        The algorithm proceeds by looking at ``a * d - b * c`` modulo
-        the ideal generated by ``G`` for some ``c`` and ``d`` with degree
-        less than ``a`` and ``b`` respectively.
-        The coefficients of ``c`` and ``d`` are indeterminates and thus
-        the coefficients of the normalform of ``a * d - b * c`` are
-        linear polynomials in these indeterminates.
-        If these linear polynomials, considered as system of
-        equations, have a nontrivial solution, then `\frac{a}{b}
-        \equiv \frac{c}{d}` modulo the ideal generated by ``G``. So,
-        by construction, the degree of ``c`` and ``d`` is less than
-        the degree of ``a`` and ``b``, so a simpler representation
-        has been found.
-        After a simpler representation has been found, the algorithm
-        tries to reduce the degree of the numerator and denominator
-        and returns the result afterwards.
-
-        As an extension, if quick=False, we look at all possible degrees such
-        that the total degree is less than *or equal to* the best current
-        solution. We retain a list of all solutions of minimal degree, and try
-        to find the best one at the end.
-        """
-        c, d = a, b
-        steps = 0
-
-        maxdeg = a.total_degree() + b.total_degree()
-        if quick:
-            bound = maxdeg - 1
-        else:
-            bound = maxdeg
-        while N + D <= bound:
-            if (N, D) in tested:
-                break
-            tested.add((N, D))
-
-            M1 = staircase(N)
-            M2 = staircase(D)
-            debug('%s / %s: %s, %s' % (N, D, M1, M2))
-
-            Cs = symbols("c:%d" % len(M1), cls=Dummy)
-            Ds = symbols("d:%d" % len(M2), cls=Dummy)
-            ng = Cs + Ds
-
-            c_hat = Poly(
-                sum([Cs[i] * M1[i] for i in range(len(M1))]), opt.gens + ng)
-            d_hat = Poly(
-                sum([Ds[i] * M2[i] for i in range(len(M2))]), opt.gens + ng)
-
-            r = reduced(a * d_hat - b * c_hat, G, opt.gens + ng,
-                        order=opt.order, polys=True)[1]
-
-            S = Poly(r, gens=opt.gens).coeffs()
-            sol = solve(S, Cs + Ds, particular=True, quick=True)
-
-            if sol and not all([s == 0 for s in sol.values()]):
-                c = c_hat.subs(sol)
-                d = d_hat.subs(sol)
-
-                # The "free" variables occuring before as parameters
-                # might still be in the substituted c, d, so set them
-                # to the value chosen before:
-                c = c.subs(dict(list(zip(Cs + Ds, [1] * (len(Cs) + len(Ds))))))
-                d = d.subs(dict(list(zip(Cs + Ds, [1] * (len(Cs) + len(Ds))))))
-
-                c = Poly(c, opt.gens)
-                d = Poly(d, opt.gens)
-                if d == 0:
-                    raise ValueError('Ideal not prime?')
-
-                allsol.append((c_hat, d_hat, S, Cs + Ds))
-                if N + D != maxdeg:
-                    allsol = [allsol[-1]]
-
-                break
-
-            steps += 1
-            N += 1
-            D += 1
-
-        if steps > 0:
-            c, d, allsol = _ratsimpmodprime(c, d, allsol, N, D - steps)
-            c, d, allsol = _ratsimpmodprime(c, d, allsol, N - steps, D)
-
-        return c, d, allsol
-
-    # preprocessing. this improves performance a bit when deg(num)
-    # and deg(denom) are large:
-    num = reduced(num, G, opt.gens, order=opt.order)[1]
-    denom = reduced(denom, G, opt.gens, order=opt.order)[1]
-
-    if polynomial:
-        return (num/denom).cancel()
-
-    c, d, allsol = _ratsimpmodprime(
-        Poly(num, opt.gens), Poly(denom, opt.gens), [])
-    if not quick and allsol:
-        debug('Looking for best minimal solution. Got: %s' % len(allsol))
-        newsol = []
-        for c_hat, d_hat, S, ng in allsol:
-            sol = solve(S, ng, particular=True, quick=False)
-            newsol.append((c_hat.subs(sol), d_hat.subs(sol)))
-        c, d = min(newsol, key=lambda x: len(x[0].terms()) + len(x[1].terms()))
-
-    if not domain.has_Field:
-        cn, c = c.clear_denoms(convert=True)
-        dn, d = d.clear_denoms(convert=True)
-    r = Rational(cn, dn)
-
-    return (c*r.q)/(d*r.p)
-
-
-def _polarify(eq, lift, pause=False):
-    from sympy import Integral
-    if eq.is_polar:
-        return eq
-    if eq.is_number and not pause:
-        return polar_lift(eq)
-    if isinstance(eq, Symbol) and not pause and lift:
-        return polar_lift(eq)
-    elif eq.is_Atom:
-        return eq
-    elif eq.is_Add:
-        r = eq.func(*[_polarify(arg, lift, pause=True) for arg in eq.args])
-        if lift:
-            return polar_lift(r)
-        return r
-    elif eq.is_Function:
-        return eq.func(*[_polarify(arg, lift, pause=False) for arg in eq.args])
-    elif isinstance(eq, Integral):
-        # Don't lift the integration variable
-        func = _polarify(eq.function, lift, pause=pause)
-        limits = []
-        for limit in eq.args[1:]:
-            var = _polarify(limit[0], lift=False, pause=pause)
-            rest = _polarify(limit[1:], lift=lift, pause=pause)
-            limits.append((var,) + rest)
-        return Integral(*((func,) + tuple(limits)))
-    else:
-        return eq.func(*[_polarify(arg, lift, pause=pause)
-                         if isinstance(arg, Expr) else arg for arg in eq.args])
-
-
-def polarify(eq, subs=True, lift=False):
-    """
-    Turn all numbers in eq into their polar equivalents (under the standard
-    choice of argument).
-
-    Note that no attempt is made to guess a formal convention of adding
-    polar numbers, expressions like 1 + x will generally not be altered.
-
-    Note also that this function does not promote exp(x) to exp_polar(x).
-
-    If ``subs`` is True, all symbols which are not already polar will be
-    substituted for polar dummies; in this case the function behaves much
-    like posify.
-
-    If ``lift`` is True, both addition statements and non-polar symbols are
-    changed to their polar_lift()ed versions.
-    Note that lift=True implies subs=False.
-
-    >>> from sympy import polarify, sin, I
-    >>> from sympy.abc import x, y
-    >>> expr = (-x)**y
-    >>> expr.expand()
-    (-x)**y
-    >>> polarify(expr)
-    ((_x*exp_polar(I*pi))**_y, {_x: x, _y: y})
-    >>> polarify(expr)[0].expand()
-    _x**_y*exp_polar(_y*I*pi)
-    >>> polarify(x, lift=True)
-    polar_lift(x)
-    >>> polarify(x*(1+y), lift=True)
-    polar_lift(x)*polar_lift(y + 1)
-
-    Adds are treated carefully:
-
-    >>> polarify(1 + sin((1 + I)*x))
-    (sin(_x*polar_lift(1 + I)) + 1, {_x: x})
-    """
-    if lift:
-        subs = False
-    eq = _polarify(sympify(eq), lift)
-    if not subs:
-        return eq
-    reps = dict([(s, Dummy(s.name, polar=True)) for s in eq.free_symbols])
-    eq = eq.subs(reps)
-    return eq, dict([(r, s) for s, r in reps.items()])
-
-
-def _unpolarify(eq, exponents_only, pause=False):
-    if isinstance(eq, bool) or eq.is_Atom:
-        return eq
-
-    if not pause:
-        if eq.func is exp_polar:
-            return exp(_unpolarify(eq.exp, exponents_only))
-        if eq.func is principal_branch and eq.args[1] == 2*pi:
-            return _unpolarify(eq.args[0], exponents_only)
-        if (
-            eq.is_Add or eq.is_Mul or eq.is_Boolean or
-            eq.is_Relational and (
-                eq.rel_op in ('==', '!=') and 0 in eq.args or
-                eq.rel_op not in ('==', '!='))
-        ):
-            return eq.func(*[_unpolarify(x, exponents_only) for x in eq.args])
-        if eq.func is polar_lift:
-            return _unpolarify(eq.args[0], exponents_only)
-
-    if eq.is_Pow:
-        expo = _unpolarify(eq.exp, exponents_only)
-        base = _unpolarify(eq.base, exponents_only,
-            not (expo.is_integer and not pause))
-        return base**expo
-
-    if eq.is_Function and getattr(eq.func, 'unbranched', False):
-        return eq.func(*[_unpolarify(x, exponents_only, exponents_only)
-            for x in eq.args])
-
-    return eq.func(*[_unpolarify(x, exponents_only, True) for x in eq.args])
-
-
-def unpolarify(eq, subs={}, exponents_only=False):
-    """
-    If p denotes the projection from the Riemann surface of the logarithm to
-    the complex line, return a simplified version eq' of `eq` such that
-    p(eq') == p(eq).
-    Also apply the substitution subs in the end. (This is a convenience, since
-    ``unpolarify``, in a certain sense, undoes polarify.)
-
-    >>> from sympy import unpolarify, polar_lift, sin, I
-    >>> unpolarify(polar_lift(I + 2))
-    2 + I
-    >>> unpolarify(sin(polar_lift(I + 7)))
-    sin(7 + I)
-    """
-    if isinstance(eq, bool):
-        return eq
-
-    eq = sympify(eq)
-    if subs != {}:
-        return unpolarify(eq.subs(subs))
-    changed = True
-    pause = False
-    if exponents_only:
-        pause = True
-    while changed:
-        changed = False
-        res = _unpolarify(eq, exponents_only, pause)
-        if res != eq:
-            changed = True
-            eq = res
-        if isinstance(res, bool):
-            return res
-    # Finally, replacing Exp(0) by 1 is always correct.
-    # So is polar_lift(0) -> 0.
-    return res.subs({exp_polar(0): 1, polar_lift(0): 0})
 
 
 def _is_sum_surds(p):
